feat: avisar que o novo produto ja existe na lista
