+++ core/database/schemas/com.z1.comparaprecos.core.database.AppDatabase/1.json
@@ -2,7 +2,7 @@
   "formatVersion": 1,
   "database": {
     "version": 1,
-    "identityHash": "32219816aa0f3f1d08b4c19d8ee251c9",
+    "identityHash": "739eec5890572208bcc9f2967ad1bae5",
     "entities": [
       {
         "tableName": "tb_lista_compra",
@@ -50,7 +50,7 @@
       },
       {
         "tableName": "tb_produtos",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `id_lista_compra` INTEGER NOT NULL, `nome_produto` TEXT NOT NULL, `quantidade` TEXT NOT NULL, `preco_unitario` REAL NOT NULL, `medida` TEXT NOT NULL)",
+        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `id_lista_compra` INTEGER NOT NULL, `nome_produto` TEXT NOT NULL, `quantidade` TEXT NOT NULL, `preco_unitario` REAL NOT NULL, `is_medida_peso` INTEGER NOT NULL)",
         "fields": [
           {
             "fieldPath": "id",
@@ -83,9 +83,9 @@
             "notNull": true
           },
           {
-            "fieldPath": "medida",
-            "columnName": "medida",
-            "affinity": "TEXT",
+            "fieldPath": "isMedidaPeso",
+            "columnName": "is_medida_peso",
+            "affinity": "INTEGER",
             "notNull": true
           }
         ],
@@ -102,7 +102,7 @@
     "views": [],
     "setupQueries": [
       "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '32219816aa0f3f1d08b4c19d8ee251c9')"
+      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '739eec5890572208bcc9f2967ad1bae5')"
     ]
   }
 }
