+++ feature/listaproduto/src/main/java/com/z1/comparaprecos/feature/listaproduto/presentation/UiState.kt
@@ -7,7 +7,8 @@ import java.math.BigDecimal
 
 data class UiState(
     val isListaProdutoCarregada: Boolean = false,
-    val produtoSelecionado: Produto? = null, //Produto(-1, -1, "", "0.0", BigDecimal.ZERO, "un"),
+    val produtoSelecionado: Produto? = null,
+    val produtoJaExiste: Produto? = null,
     val listaCompra: ListaCompra = ListaCompra(-1, "", false, -2, 0L),
     val listaProduto: List<Produto> = emptyList(),
     val listaCompraComparada: ListaCompraWithProdutos = ListaCompraWithProdutos(listaCompra, emptyList())
